Fix xrWeatherEditor compilation and compiler warnings

+++ src/editors/xrWeatherEditor/entry_point.cpp
@@ -34,7 +34,7 @@ protected:
     {
         LONG_PTR result;
 #ifdef XR_X64
-        if (m_engine && m_engine->on_message((HWND)m.HWnd.ToInt32(), m.Msg, m.WParam.ToInt64(), m.LParam.ToInt64(), result))
+        if (m_engine && m_engine->on_message((HWND)m.HWnd.ToInt64(), m.Msg, m.WParam.ToInt64(), m.LParam.ToInt64(), result))
             return;
 #else
         if (m_engine && m_engine->on_message((HWND)m.HWnd.ToInt32(), m.Msg, m.WParam.ToInt32(), m.LParam.ToInt32(), result))
